Complete re-factoring.
Removed dependencies on external modules. This consisted mainly of 'apt',
and 'common, which 'apt' depends on. Created an optional drush::apt class
that will provide the minimum for apt to function, but can be turned off
with a parameter to the main drush class. This makes the installation itself
optional, and should allow for better portability.

Made heavy use of a drush::defaults class to provide default arguments to
all classes and defines. Defined a general drush::run resource type to allow
running Drush commands easily and consistently. Rebuilt other resources to
wrap around drush::run.

Updated docs to describe the current functionality in lots of detail.